Adding remote test net model  (#1)
* Create appRemote.py

* Create .gitignore

* Update README.md

* fixing pip and npm issues

* few fixes
