+++ .gitignore
@@ -1,0 +1,19 @@
+# Build and Release Folders
+bin-debug/
+bin-release/
+[Oo]bj/
+[Bb]in/
+
+# Other files and folders
+.settings/
+node_modules/
+build/
+package-lock.json
+__pycache__/
+# Executables
+*.swf
+*.air
+*.ipa
+*.apk
+.env
+*.env
