+++ README.md
@@ -1,5 +1,26 @@
-# IoT-and-Blockchain
+## IoT-and-Blockchain
 
-follow this medium article for in instructions https://link.medium.com/TcNvHHZoY1 
+- [Demonstration Article](https://link.medium.com/TcNvHHZoY1)
 
- © Salman Dabbakuti
+##### Quick Demo
+```
+git clone https://github.com/Salmandabbakuti/IoT-and-Blockchain.git
+cd IoT-and-Blockchain
+npm install
+ganache-cli # local testnet for deployments
+truffle compile 
+pip install -r requirements.txt
+python app.py # Headover to your Computer IP on browser for UI
+
+python appRemote.py # Needs to add your private key inorder to work.
+```
+##### UI:
+
+<img align="center" src="https://github.com/Salmandabbakuti/IoT-and-Blockchain/blob/master/assets/screen.png" width="90%">
+
+
+ ##### Author   
+
+##### :wave: [Salman Dabbakuti](https://salmandabbakuti.github.io)
+
+<a href="https://www.buymeacoffee.com/Salmandabbakuti" target="_blank"><img src="https://www.buymeacoffee.com/assets/img/custom_images/orange_img.png" alt="Buy Me A Coffee" style="height: 41px !important;width: 174px !important;box-shadow: 0px 3px 2px 0px rgba(190, 190, 190, 0.5) !important;-webkit-box-shadow: 0px 3px 2px 0px rgba(190, 190, 190, 0.5) !important;" ></a>
