+++ package.json
@@ -1,0 +1,19 @@
+{
+  "name": "iot-blockchain-demo",
+  "version": "1.0.0",
+  "description": "First implementation of enhanced security in IoT using Blockchain",
+  "main": "",
+  "scripts": {
+    "build": "npm install",
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "complie": "truffle compile",
+    "start": "ganche-cli"
+  },
+  "author": "Salman Dabbakuti",
+  "license": "ISC",
+  "dependencies": {
+    "ganache-cli": "^6.9.1",
+    "pip": "0.0.1",
+    "truffle": "^4.1.17"
+  }
+}
